test: remove tests for deprecated `hook.remove.{before, error, after, wrap}()` methods
