+++ test/integration/constructor-test.js
@@ -8,8 +8,6 @@ test('Constructor', function (t) {
   t.is(typeof hook, 'function', 'hook() is a function')
   t.is(typeof hook.before, 'function', 'hook.before() is function')
   t.is(typeof hook.after, 'function', 'hook.after() is function')
-  t.is(typeof hook.remove.before, 'function', 'hook.remove.before() is function')
-  t.is(typeof hook.remove.after, 'function', 'hook.remove.after() is function')
 
   t.is(typeof hook.api, 'object', 'hook.api is an object')
 
@@ -21,8 +19,6 @@ test('hook.api', function (t) {
 
   t.is(typeof hook.api.before, 'function', 'hook.before() is function')
   t.is(typeof hook.api.after, 'function', 'hook.after() is function')
-  t.is(typeof hook.api.remove.before, 'function', 'hook.remove.before() is function')
-  t.is(typeof hook.api.remove.after, 'function', 'hook.remove.after() is function')
   t.is(typeof hook.api, 'object', 'does ont expose hook() method')
 
   t.end()
