+++ test/integration/index.js
@@ -4,7 +4,5 @@ require('./before-test')
 require('./error-test')
 require('./hook-test')
 require('./singular-hook-test')
-require('./remove-after-test')
-require('./remove-before-test')
-require('./remove-error-test')
 require('./wrap-test')
+require('./remove-test')
